Split the renamespacer into bits that could be reused.
